refactor: extract agent cards from ChatProvider into direct React Query consumers
Removes agent-data fields (agents, agentsError, agentsLoading, agentsRefetch,
agentNameDisplayNameMap) from ChatContext. Consumers now call useAgentCards()
directly, letting React Query handle caching and deduplication. Adds
staleTime: 30s so navigating between tabs triggers background refetch.
Exports agent-cards module from the package for enterprise reuse.
