+++ client/webui/frontend/src/lib/api/index.ts
@@ -1,4 +1,5 @@
 export { api, getErrorFromResponse, refreshToken, scheduleProactiveRefresh, cancelProactiveRefresh } from "./client";
+export * from "./agent-cards";
 export * from "./features";
 export * from "./people";
 export * from "./projects";
