+++ client/webui/frontend/src/lib/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from "./useAgentCards";
 export * from "./useArtifactOperations";
 export * from "./useArtifactPreview";
 export * from "./useArtifactRendering";
