--- client/webui/frontend/src/lib/hooks/useAgentCards.ts
@@ -1,6 +1,0 @@
-/**
- * Re-export from the new React Query-based module.
- * Kept for backwards compatibility with existing imports.
- */
-export { useAgentCards } from "@/lib/api/agent-cards/hooks";
-export { transformAgentCard } from "@/lib/api/agent-cards/service";
